requirements.txt
tensorflow>=2.8.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
Pillow>=8.3.0
opencv-python>=4.5.0
joblib>=1.1.0